// crates/clip-mash-app/src/api.ts
export interface SongDto {
  beats: number[]
  duration: number
  fileName: string
  songId: number
  url: string
}